requirement.txt
flask
sqlalchemy
mysql-connector-python
Flask-SQLalchemy
Flask-script
Flask-migrate
Flask-Session
Flask-WTF
captcha
Flask-Bcrypt
bcrypt
Flask-Mail
Flask-Login
Flask-OpenID
Flask-RESTful
Flask-Bootstrap
Flask-Moment
Flask-Admin
Flask-RESTX
Flask-Principal
Flask-Security
Flask-marshmallow
Flask-PyMongo
Flask-Caching
Flask-DebugToolbar
Flask-Assets
Flask-Uploads
Flask-Cors
Flask-SocketIO
flask_jwt_extended
Flask-sse
jwt
pyjwt
requests
lxml
bs4
python-dotenv
argon2-cffi
apscheduler
openai